Add support for pulling authorities from a "roles" claim
Currently to assign roles to a user within jenkins via keycloak the
role must be placed at a realm global level and full scope allowed
must be on.

This isn't idea.... it would be nice to be able to use client roles.
This commit adds support for reading roles from a claims attribute
called "roles".

A keycloak administrator can associate roles to a user using the
mappers in keycloak